+++ src/InformaticsGateway/Services/Scp/ApplicationEntityManager.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 MONAI Consortium
+ * Copyright 2021-2023 MONAI Consortium
  * Copyright 2019-2021 NVIDIA Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -171,6 +171,25 @@ namespace Monai.Deploy.InformaticsGateway.Services.Scp
             }
         }
 
+        private void UpdateAeTitle(MonaiApplicationEntity applicationEntity)
+        {
+            if (_aeTitles.TryGetValue(applicationEntity.AeTitle, out var aeHandler))
+            {
+                try
+                {
+                    aeHandler.Configure(applicationEntity, Configuration.Value.Dicom.WriteDicomJson, Configuration.Value.Dicom.ValidateDicomOnSerialization);
+                }
+                catch (Exception ex)
+                {
+                    _logger.FailedToUpdateAppliationEntityHandlerWithUpdatedAEChange(applicationEntity.AeTitle, ex);
+                }
+            }
+            else
+            {
+                _logger.FailedToUpdateAppliationEntityHandlerWithUpdatedAEChange(applicationEntity.AeTitle);
+            }
+        }
+
         public void OnCompleted()
         {
             // noop
@@ -189,6 +208,10 @@ namespace Monai.Deploy.InformaticsGateway.Services.Scp
             {
                 case ChangedEventType.Added:
                     AddNewAeTitle(applicationChangedEvent.ApplicationEntity);
+                    break;
+
+                case ChangedEventType.Updated:
+                    UpdateAeTitle(applicationChangedEvent.ApplicationEntity);
                     break;
 
                 case ChangedEventType.Deleted: